KF-CuP/radiation and KF-CuP/cloud interactions
Adds cldfra_dp and cldfra_sh outputs from KF-CuP, the cloud fractions
associated with deep and shalow convection. Calculations are based on
how cldfra_dp and cldfra_sh are already calculated in kfeta since
WRF3.6.
cldfra_dp and cldfra_sh are already passed from the cumulus driver
(from e.g., KFETA) to the radiation driver, where they are added to the
total cloud fraction, CLDFRA, and used in the radiation calculations.
This commit adds a condition in phys/module_physics_init.F allowing this
to happen for KF-CuP also, when icloud=1 is selected in the namelist.