Fix chem initialization bugs in MOZART and MOSAIC
Add initializations for several arrays in MOZART and MOSAIC in chem, in
the aerosol chemistry code, in the profile initialization for chemical
variables in real, and in the irr_diag reaction rate output.
These issues trigger crashes at runtime at the model initialization
stage when using debug compilation flags, and were preventing the use of
debug options for investigating other bugs.
In addition, when using gfortran, these arrays would be be initialized
with arbitrary values, then used in the chemistry, causing some aerosol
output to diverge to 0 or infinity.

A fix is applied to initialize a diagnostic array "irr_rates" for all chem_opt.
A "not allocated" error (when compiled with a debug option) is avoided
not only for T1_MOZCART_KPP, MOZCART_KPP, MOZART_MOSAIC_4BIN_AQ_KPP, MOZART_MOSAIC_4BIN_KPP
but also all other chemistry options (chem_opt).